Added sorting
In the client database viewer you can now sort every entry by the column value by clicking the column title. Had to change the system for
the editing of the accounts list according to the row selected by the treeview.

And some additional minor fixes and minor code layout changes to the client and server.

Due to the new complexity of the system I can't really be bothered to redesign for searching, so this hopefully is the last change assuming everything is functional.